pyproject.toml
[build-system]
requires = ["setuptools>=71.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ml-production-template"
version = "0.1.0"
description = "A production-ready ML template with MLflow, Prefect, and FastAPI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    "scikit-learn>=1.4",
    "pandas>=2.0",
    "numpy>=1.26",
    "mlflow>=2.12",
    "prefect>=2.19",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "imbalanced-learn>=0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 88